+++ changelog.md
@@ -22,6 +22,7 @@ Codion Change Log
 - EntityEditPanel.beforeInsert(), beforeUpdate() and beforeDelete() removed.
 - EntityTablePanel bug fixed, no longer creates a condition panel if one should not be included.
 - EntityEditPanel.confirmInsert(), confirmUpdate() and confirmDelete() now protected, useful when doing custom crud operations.
+- EntityTablePanel.confirmDelete() added.
 
 ## 0.17.21
 ### is.codion.swing.common.model